Give the name grammar and the entry reader their own modules
Both modules the variants work has to grow sit exactly on the 500-line budget, so
before adding anything to them this change takes two pieces out that were never
about the module they lived in. The identifier and reference grammar moves from
model.py into names.py, where the loader and the schema exporter reach for it
without going through the models; the model module's own docstring already
admitted the grammar was a guest there.

Reading one component entry moves from loader.py into entries.py. An entry is the
same block wherever it is written — at the top level, inside a group and, from the
next commit on, inside a variant option — so the code that reads one belongs apart
from the code that reads the document around it. The loader keeps the document: the
schema version, the top-level keys and the blocks that hold components.

Nothing changes behaviourally. The methods that moved were private to the reader,
the four input-item key constants moved with them, and the two public spellings
(EnergySystemReader, NameRules re-exported from the package) are untouched.
